l10n_br_fiscal: merge tax_definition classes
