+++ l10n_br_fiscal/models/__init__.py
@@ -36,12 +36,6 @@ from . import tax_ipi_guideline
 from . import tax_ipi_guideline_class
 from . import tax_estimate
 from . import tax_definition
-from . import tax_definition_benefit
-from . import tax_definition_company
-from . import tax_definition_cfop
-from . import tax_definition_icms
-from . import tax_definition_operation_line
-from . import tax_definition_partner_profile
 from . import icms_regulation
 from . import icms_relief
 from . import document_type
